_UnternehmensDjango/ix20-12-120_DjangoUnternehmen/202010_django/ix_django-master/requirements/base.txt
asgiref==3.2.10
Django==3.1
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.9.2
django-filter==2.3.0
django-python3-ldap==0.11.3
django-tables2==2.3.1
djangorestframework==3.11.1
ldap3==2.8
Markdown==3.2.2
pyasn1==0.4.8
pytz==2020.1
sqlparse==0.3.1

